+++ lib/exe.js
@@ -7,7 +7,7 @@ const tweetlog = require('../lib/tweetlog')
 const log = tweetlog('n3h-bin')
 
 const { buildLogHandler } = require('../lib/file-logger')
-const { N3hHackMode, N3hMock } = require('../lib/index')
+const { N3hHackMode, N3hMock, N3hRealMode } = require('../lib/index')
 const { unhandledRejection } = require('../lib/n3h-common')
 unhandledRejection.strict()
 
@@ -114,6 +114,9 @@ async function main () {
 
   log.i('executing mode ' + mode)
   switch (mode) {
+    case 'REAL':
+      n3hNode = await new N3hRealMode(workDir, rawConfigData)
+      break
     case 'MOCK':
       n3hNode = await new N3hMock(workDir, rawConfigData)
       break
@@ -121,7 +124,7 @@ async function main () {
       n3hNode = await new N3hHackMode(workDir, rawConfigData)
       break
     default:
-      log.e('so called "real" mode disabled while running POC on hackmode / mockmode')
+      log.e('so called "real" mode disabled while running POC on realmode / mockmode')
       process.exit(1)
       // n3hNode = await N3hNode.constructDefault(workDir, rawConfigData)
       // break
