+++ lib/index.js
@@ -13,6 +13,7 @@ const DEFAULT_MODULES = [
 ]
 */
 
+exports.N3hRealMode = require('./realmode').N3hRealMode
 exports.N3hHackMode = require('./hackmode').N3hHackMode
 exports.N3hMock = require('./n3h-mock').N3hMock
