+++ lib/n3h-ipc/n3hMode.js
@@ -5,7 +5,7 @@ const { AsyncClass } = require('../n3h-common')
 const { Connection } = require('../n3h-mod-spec')
 const { ConnectionBackendWss } = require('../n3h-mod-connection-wss')
 
-const config = require('../hackmode/config')
+const config = require('../realmode/config')
 
 const tweetlog = require('../tweetlog')
 const log = tweetlog('*n3hMode*')
