+++ lib/realmode/index.js
@@ -8,7 +8,7 @@ const { N3hMode } = require('../n3h-ipc/n3hMode')
 const { Mem, getHash } = require('./mem')
 
 const tweetlog = require('../tweetlog')
-const log = tweetlog('@hackmode@')
+const log = tweetlog('@realmode@')
 
 const { DhtBackendFullsync } = require('../n3h-mod-dht-fullsync')
 
@@ -18,7 +18,7 @@ const {
 } = require('../n3h-mod-spec')
 
 /**
- * N3h "hackmode" prototyping code
+ * N3h "realmode" prototyping code
  *
  * Expects a config either over stdin or as a file `n3h-config.json` in the
  * working directory.
@@ -44,7 +44,7 @@ const {
  * - `webproxy.wssAdvertise` {uri|"auto"} - Cannot be paired with `wssRelayPeers`. Sets up this node to be directly connectable at this address. Special case if set to `"auto"` will pick the first public NIC binding... allowing for os-assigned ports.
  * - `webproxy.wssRelayPeers` {array<uri>} - Cannot be paired with `wssAdvertise`. Uri array of relay peers to connect through. (currently you can only specify 1). Use this if behind a NAT, all messages will be routed through the peer specified here.
  */
-class N3hHackMode extends N3hMode {
+class N3hRealMode extends N3hMode {
   async init (workDir, rawConfigData) {
     await super.init()
 
@@ -1080,4 +1080,4 @@ class N3hHackMode extends N3hMode {
   }
 }
 
-exports.N3hHackMode = N3hHackMode
+exports.N3hRealMode = N3hRealMode
