+++ lib/realmode/index.test.js
@@ -2,8 +2,8 @@ const { expect } = require('chai')
 
 const lib = require('./index')
 
-describe('HackMode Suite', () => {
+describe('RealMode Suite', () => {
   it('should be a function', () => {
-    expect(typeof lib.N3hHackMode).equals('function')
+    expect(typeof lib.N3hRealMode).equals('function')
   })
 })
