+++ lib/realmode/mem.js
