+++ lib/realmode/p2p-backend-hackmode-peer.js
@@ -2,7 +2,7 @@ const { AsyncClass, Track } = require('../n3h-common')
 const { Keypair } = require('../hc-dpki')
 const mosodium = require('../mosodium')
 const tweetlog = require('../tweetlog')
-const log = tweetlog('p2p-hackmode')
+const log = tweetlog('p2p-realmode')
 
 const { URL } = require('url')
 const msgpack = require('msgpack-lite')
