+++ lib/realmode/p2p-backend-hackmode-peer.test.js
@@ -12,7 +12,7 @@ const { expect } = require('chai')
 const { P2p } = require('../n3h-mod-spec')
 const { P2pBackendHackmodePeer } = require('./p2p-backend-hackmode-peer')
 
-describe('hackmode module p2p peer backend Suite', () => {
+describe('realmode module p2p peer backend Suite', () => {
   let allNodes = []
 
   const dht = JSON.stringify({})
