+++ lib/realmode/util58.js
