+++ lib/realmode/util58.test.js
